rename simple_client_test file to exclude it from pytest
